Made some pretty fundamental changes to walker to ensure that walkers always stop walking at the exact positions they were told to.
Up to this point, I fudged the end walk point a bit. When the walker was close enough to the destination, I'd move them back slightly so that the "end walk" animation would get you mostly in the correct spot. However, I often encountered bugs related to walkers not stopping in exactly the right spots. The original game seemed to do this correctly of course.

On closer inspection, I noticed a big oversight: the original game doesn't even USE the end walk animations! Via some testing, I verified that this was the case. As a result, I updated my code to also not use the "end walk" animation, and ensured that the walker always gets set to the exact end position when the walk ends.

This still needs to be tested thoroughly...anticipate bugs until a full game walkthrough is done successfully.