Fixed a potential issue where Config File "GetX" functions could get mixed up due to implicit conversion rules.
When attempting to use the Config::GetBool(string, string, bool) function, I noticed that the compiler was actually using the (string, bool) version because it would implicitly convert the string to a boolean! This is very silly to me, but to be on the safe side, I modified the functions to no longer take default parameters.