Implemented first pass at pausing the game and audio when the application window is not focused.
In most games, including the original GK3, if the app window loses focus (perhaps you switch to a browser or another app), the game and audio pause until you return to the game. Up to this point, that wasn't the case with this engine.

This commit adds support for that as the default behavior in release mode (debug mode still stays active in the background - it's useful for debugging). There's also a new config flag to control this behavior in release too.