Maybe(?) fixed occasionally unresponsive IMGUI tools menu.
I'm not 100% sure about this fix, but I noticed that if the game was running for a decent amount of time, the IMGUI tools menu would be unresponsive for awhile. I'm no IMGUI expert, but my hunch is that even if the tool is inactive, we should be calling the IMGUI render functions every frame.

After making this change, I haven't seen the problem again - fingers crossed!