Fixed hiding action bar by right-clicking not triggering the cancel callback.
When you hide the action bar by pressing the "cancel" button, it triggers some code that notably cancels the current conversation if any. However, when the action bar was hidden due to a right-click or other scenarios, the cancel code wasn't being executed, which could leave you in an odd situation (such as still playing talk/listen anims when no longer talking to someone).

This commit fixes that by adding a way to call the "cancel" callback when hiding the action bar.