Fixed two men entering car in Day 1, 6PM being glitchy.
At some point, I made the assumption that if you toggle a submesh's visibility, you must want the entire mesh to be enabled, right? Turns out that's not true, and doing that causes some 3D models to appear at incorrect times. Such as when the two men enter their car in this timeblock.