Fixed a soundtrack with no playable nodes continuing to play indefinitely.
Most soundtracks in the game loop indefinitely, but there are a select few, for one-off sound effects, that are meant to stop playing after some time. The indication of this seems to be if all nodes in the soundtrack have a repeat limit specified, and they all have repeated the max amount. So, added some code to detect this and stop the soundtrack in that case.

In particular, this fixes a bug where rotating the laser heads in Day 2, 12PM would only play the "laser rotate" SFX a single time.