Fixed Gabriel's walk behavior during mid timeblock cutscene in Day 2, 5PM.
Right in the middle of this timeblock, Grace can go down to the Lobby to see a cutscene where Gabe, Mosely, and Buthane all go out to the bar. Gabe's behavior during this cutscene was very erratic - walking in circles, teleporting to opening the door, etc.

Upon closer inspection, the script for this cutscene relies on the idea that if a character is walking, and some other animation plays on the character, the walk should cancel. This is the first (only?) spot I've seen in the game where this behavior is expected. So what would happen is Gabe would be walking to the lobby door, but then was told to play a "door open" anim during the walk. The game gets confused and warps Gabe to open the door, but then also tries to continue the walk - multiple animations try to play over one another and chaos ensues.

To fix this, I saw no choice but to detect when a non-walk animation plays during a walk - and stop the walk in that scenario. After making this change, the cutscene works as expected.