Fixed skybox not rendering at Chateau de Blanchforte parking lot during Day 3, 7AM timeblock.
In OpenGL, skyboxes are required to be square and all sides must have the same resolution. For this one scene, there's an asset bug - one of the skybox sides is smaller than the others. So the skybox doesn't render.

To fix this, I'm checking for any size that doesn't match and upscaling to force it to match.