Fixed a soft lock when talking to Mosely in Dining Room during Day 3, 7AM (due to a data entry error).
In a couple cases, GK3 animations have data entry errors - an animation node specifies an out of bounds frame number, meaning that the node will never be executed. In this case, the animation is only 11 frames long, but the node is set to execute on frame 98! Must be a typo or oversight!

To fix this, I now just clamp animation node frame numbers within the valid range.