Fixed "demon breathing" SFX continuing to play over closing cutscene and credits.
Upon investigating the demon breath SFX are a soundtrack file, and they are played using the mysterious "PlaySoundtrackTBS" animation command. What is the "TBS" and how does it differ from the normal "PlaySoundtrack" command? Nothing in the docs explain the difference.

However, I did some tests in the original game with customized assets and found that the "TBS" variant appears to only play the soundtrack once, with no looping. This wasn't noticed with other "TBS" style soundtracks because all others (besides this demon breathing one) also define repeat limits that cause them to be de-facto non-looping anyways.