Improved Gabe's behavior when fighting Demon in last scene of Day 3, 9PM timeblock.
Like the other rooms in the temple, Gabe's behavior in this final scene is greatly limited. He isn't allowed to walk freely, and attempting to walk performs a different action. Additionally, Gabe's "walk back" anims were causing a lot of trouble, so I brute forced a solution to make it look better.