Removed walker override anim behavior, which was no longer necessary and was causing a turn-to-face bug.
In the past, some logic was adding to the walker to have it behave slightly differently when using override walk anims. However, this logic is no longer needed, since we aren't using walk end anims anymore. Removing this logic also fixes a bug where "turn to face" behavior at the end of a walk was being ignored when using an override walk anim.