Added a hack solution to stop Prince James from falling into the floor when you meet him in the lobby during Day 3, 12PM timeblock.
The animation system in this game continues to be frustrating - in this case, logic that works for every other character in the game causes this one character to position incorrectly. A hack is to just exclude this character from starting fidgets on init. Eventually, it'd be good to go back and review the animation system in totality to see if this can be improved upon...