Adjusted the near clipping plane to be a bit farther away, to avoid clipping in a few instances.
The near clip plane is a delicate balance between avoiding clipping with things close to the camera and clipping with things far from the camera.

In this case, using a near clip plane value of 1 was too strict. In at least one case (doing the tilted square with the angels) the position of Gabe and the camera resulted in Gabe's hair being visible on screen when it shouldn't be. Increasing the near clip plane to 12 (which the original game used) fixes that issue.

However, using a value of 12 causes some unideal clipping with scene geometry as you navigate close to walls (this is even present in the original game). As a happy medium (I hope), a value of 8 seems to avoid both problems. After testing the majority of the game using a value of 8, I didn't notice any big problems.