Improved "walk to see" detection of the target object by raycasting against additional AABB points.
Previously, to detect if a walking character could see the target object, we only did a single raycast against the center of the object's AABB. However, this is not sufficient for objects that might be partially obscured. A good example is Jean/Simone standing at the desk in the Lobby. Though they're clearly visible to the player character, the raycast from the character head to the AABB center hits the desk and the game thinks we can't see them.

To fix this, I am just doing two additional raycasts to the AABB's min and max points. This should provide pretty good coverage of whether an object can be seen at a particular spot.