Fixed walk skip with a "walk to see" target defined always skipping to the end of the path, rather than the first spot along the path that the target becomes visible.
When walking normally with a "walk to see" target, the character stops walking as soon as the target object comes into view. Consider "looking" at the couch in Room 25: Gabe/Grace don't need to walk right up to it to comment on it, it just needs to come into view.

An oversight in the "skip walk" code would always put the character at the final path position, even when using a "walk to see" target. The effect of this was some pretty awkward behavior - Gabe/Grace would stand immediately next to the target object after the skip. To fix this, we now check each point along the path during a "skip walk" request to see if the object should be in view at that spot. If so, we skip to that spot instead of the end point of the path.