Fixed (I think) bug where spamming walk actions could result in character becoming frozen and not walking anywhere until you hit the ESC key.
If you click to walk somewhere when a walk is already in progress, the walk system tries to simply reroute the current walk to the new destination. This looks pretty natural. However, when this happens, we need to be careful about ensuring state variables related to continuing walk animations stay in the correct state - or else walk anims might stop and the character becomes stuck, unable to continue moving. Fortunately, skipping the walk still works as a fallback.