Refactored video playback code to use a "VideoImage" component instead of a global manager class.
Video player code was originally implemented as a global manager class that allows playback of a single movie through a few UI elements. The manager would manually update the pixel data and ensure the texture was being displayed on a UI image. This worked OK, but when the need arose to play videos on arbitrary UI textures (for Sidney), this architecture didn't quite work. The abilty to play the video on arbitrary textures was very much "hacked in."

Recently, I wanted to fix a bug where a single frame of the 3D scene would render between timeblock end movies and the start of the next timeblock. It was proving difficult to do with the current approach, so I bit the bullet and did a bit of refactoring.

Now, there is a "UIVideoImage" component that acts like a UI image, but can play a video via the public API. This component can be used ANYWHERE in UI, so it more elegantly solves the problem of playing videos in Sidney (the Sidney screens that need to play video just create "UIVideoImage" components and configure them as needed).

For the global "video player," this also allowed the code to be simplified significantly. The video player is now just another UI screen that uses a UIVideoImage to display a video (with some extra logic related to sizing/scaling the movie).