Added "Color" class and made some improvements to "Color32."
The "Color32" class was added early in engine development, so I took a pass to clean up some of the older code in there. I even fixed a few bugs in there, such as needlessly converting rgba to floats when checking equality and accidentally multipliying the "green" component twice in the *= operator override.

Additionally, I didn't like having to manually convert colors between byte and float representation, so I finally introduced a "Color" class with implicit conversion to/from Color32. "Color" differs from "Color32" in that it uses floats to represent "percentage of each color component." It doesn't clamp values, so it can represent colors outside the normal range of 0-1. It can be convenient to use when passing colors to be used by graphics systems or when needing to convert between the two formats.