Improved performance of texture loader for large BMP image assets.
In response to a bug report that the hi-res texture pack loaded very slowly, I investigated why that might be the case. The BMP loading code was reading one pixel at a time, and doing a manual BGR to RGB conversion at the same time. It turns out this is really slow, especially with a "hi-res" texture pack where all the texture resolutions have increased significantly!

To fix this, I had to make some pretty huge changes to the texture class - in particular, supporting a few different texture formats rather than forcing every texture to RGBA format. This allows many textures to be read in using a single read statement, which is very efficient compared to the old method. The result is that "hi-res" texture loading performance doesn't feel any slower than normal texture loading (though it technically does take a few more milliseconds due to the increased number of pixels).

This was a fairly complex change - file it under "I hope that didn't break anything!"