Added support for combining vertex/fragment shaders into single file. Added support for shader feature flags. Consolidated all shaders into a single "uber" shader.
Before this change, G-Engine had 12 separate shader source files. Over time, I've found this difficult and confusing to maintain, especially when a piece of shader code needs to be duplicated across several files. Beyond the annoyance of maintaining duplicate code, this becomes error-prone if you change one file and forget to change the others. Recently, when I found myself copy/pasting the "discard color" behavior across all 5 fragment shader source files, I decided something needed to be done to improve the situation.

To do this, when loading and compiling a shader, it's now possible to specify feature flags. In the shader, this becomes a #define for each given flag. The shader source can then use an #if with the feature flag to conditionally include code from the shader source. This allows many different specific shaders to be compiled from one vertex or fragment shader source. Any shared source code is simply not inside a #if, so it always gets compiled.

Building off of this further, I added support for putting both vertex and fragment shaders in a single source file. This is achieved by adding a #define of VERTEX_SHADER when compiling the vertex shader, and FRAGMENT_SHADER when compiling the fragment shader. The source file then just needs to ensure that the vertex and fragment parts are within those #ifs. This is totally optional, and it's still possible to put vertex/fragment sources in separate files if desired.

One final benefit of this change is to require shader source files to use the conventional extension for the graphics API (.glsl for OpenGL). Since the extension is now retrieved from the graphics API, this helps pave the way for adding another graphics system in the future.