Implemented UI scaling logic, which should make the game more playable at very high resolutions. Very WIP and experimental right now!
Up to this point, the majority of UIs in the game always rendered at their originally authored size of 640x480. At resolutions higher than that, the UI would simply become smaller. This works well and looks nice up to a point, but at high resolutions like 4K, the UI is so small that the game becomes unplayable.

To alleviate that, this commit introduces logic to scale UIs 2x, 3x, 4x or more depending on the current game's resolution relative to the authored resolution. In particular, the window height is used - if the window height is greater than 480x2, the UI is scaled 2x. If greater than 480x3, the UI is scaled 3x. And so on.

There was a decent amount of refactoring work required to ensure existing UIs weren't assuming they were at 1x scale - particularly when dragging or mouse position was involved. But fortunately the scaling method used means the majority of UI code didn't have to change.

This change temporarily removes the custom scaling logic used on some screens (title screen and driving screen to name a few). I'm planning to re-introduce that logic, along with some UI scaling preferences/settings, in future commits.