Added a new "Settings Tool" that uses IMGUI to expose additional game settings that weren't included in the original game. WIP on defining UI scaling settings.
The original game exposed a variety of audio/video/game settings. But as features are added to this engine rebuild, additional settings may be needed. Adding those settings to the original in-game menu is problematic for two reasons. First, the original menu was not built in an easily extendable way - the only way to add more settings is to update art. Second, adding more settings detracts from the original game's experience.

To balance maintaining the original experience while also providing a way to show more settings, a special IMGUI-based menu is introduced that can expose such settings. This menu is accessible via the "tab" key when running the game. It's also accessible via a new button that appears when you enter the "Advanced Options" section of the Options Bar.