Fixed (I think?) syncing video playback to audio not working correctly.
Way back when I first implemented video playback, I couldn't get "sync video to audio" behavior to work - the video would become very choppy. Feeling a bit in a rush, I noticed syncing to external clock seemed to work fine, so I just went ahead and used that setting instead.

However, syncing to audio is, I think, the ideal approach. Syncing to an external clock can introduce some delay between video/audio (this is most noticeable during the Sierra logo video, which is slightly desynced). And syncing to video can cause audio delays which are noticeable. But syncing to audio just means that the video system might display a frame longer to slow down or skip a frame to catch up.

After a bit of head scratching, I think I found the reason audio sync wasn't working. There's a spot in the video decoder with logic to decide whether to skip a video frame. A condition had been commented out because I thought it wasn't relevant (it was from part of the ffmpeg sample code that I didn't use, related to video filters). However, a simplified version of the condition is actually necessary - after adding this back, the audio sync mode seems to work correctly. And the Sierra video is synced correctly!