Improved "Skip to View" logic under certain edge cases.
One improvement here is to use the "turn to face" direction when skipping to the last node of a walk path. This causes characters to appear at the last node and facing the right way. Previously, there were some awkward cases where the character would appear at the final spot, but facing the wrong way - they'd then have to turn on the spot to the target direction.

Another big improvement is to make the logic for checking whether a walk position is on screen or not more comprehensive. Previously, it would only check the walk path position and the estimated position of the character's head. The updated logic checks a line segment between those two points, which is more likely to accurately detect if the character would be visible on-screen at that position.