Revised "walk to see" behavior to be precalculated, rather than checked each frame.
Despite best efforts, I was still seeing some scenarios where the "walk to see" behavior was failing. The problem was only checking AABB raycasts for whether a thing was seen, but not considering if an AABB was completely obscured by a larger AABB. In this case, the character would act like they couldn't ever see the object.

This commit changes the logic to do a normal raycast (so checks for hitting objects, not AABBs) but then critically use rays based on the target object's AABB *and* consider an object "seen" if the raycast hits the target object OR if the target object's AABB is closer than whatever was hit. This seems to give good results, but as always, time will tell and a lot of cases need to be tested.

Along with that change, I also realized that I could precalculate the walk to see spot during initial path calculation, rather than checking it every frame. The main benefit here is improving performance (doing less work each frame, and many raycasts can start to bog things down) and streamlining logic (both skip to end and normal pathing can use the same logic now). The only downside I can see is if the walk to see target is moving, but that is rare (or possibly non-existent in this game), and the effect would be pretty minor.