Added "walk to see" goal position bias based on target object facing direction.
One problem I noticed in the Museum was, if you "looked" at the panels facing away from the entrance, the character would walk right up to the panel from behind and stop there. The problem is that the character paths directly to the panel spot (which walks up behind it), never sees the panel, and then ends up in a weird ending spot.

To help fix this, and to act as a hint, I am now using the facing direction as a hint when calculating the path to a "walk to see" object. The final destination point is offset a little bit in the facing direction of the object. In the Museum case, this causes the character to intelligently walk around the panel and look at it from the front side.

As part of this, I also had to add calculation of facing direction for BSP Actors. Until now, they always had the default rotation. But now, the facing direction is averaged from all the surfaces/polygons in the object and used to calculate a heading.