Walk skip no longer occurs if it would skip a trigger region along the walk path.
GK3 allows walk paths to be shortened dramatically. If 90% of the walk path is off-screen, that 90% will be skipped, teleporting the character to the closest offscreen path node. If ESC is pressed, the entire walk path can be skipped, teleporting the player character to the end point immediately.

However, there's a problem here: trigger regions. What if a walk path would have caused the player character to enter a trigger region (causing the trigger action to execute)? The skip behavior could miss the trigger - that's a bug.

To fix this, code that shortens a path now checks whether the path intersects any active trigger region in the level. If so, the walk path is not allowed to skip past that point. The result is that you can't avoid hitting the trigger if the path would have taken you through there, even if you use a skip technique.