Added logic to stop walking if position is set via script.
This seems to math the original game's behavior. If walking and the actor position is changed (aka a teleport), the walk should be cancelled/stopped.

This fixes at least one bug: at Devil's Armchair in Day 2, 2pm, Gabe would overshoot his mark when walking up to the chair. This change resolves that.