Separated scene load and init from scene enter logic.
In anticipation of properly supporting scene save/load logic, this commit separates scene load/init from scene enter code. This is important because "scene enter" logic is only meant to be called when you enter the scene from some other scene in the game - going from hallway to R25 for example. When loading a save, the enter logic shouldn't occur - you were already in the room, so you did not "enter" from another location when loading a save.