Added initial pass at scene save/load state logic.
One big missing piece of the current save/load code was that scene state was not included. If you loaded a save game, the game would always act as though you had just entered that scene fresh. Things like camera state and the state of scene actors were not remembered. In many cases this was merely inconvenient, but there were also times where it could soft-lock the game or cause bugs.

So, this commit adds the basics of saving and loading scene state. It seems to work reasonably well, but I must admit I'm very unsure about this approach long-term. We'll see how it goes. Worst-case, a better solution can be implemented later, though maintaining backwards compatibility of saves would be a pain.