WIP on save/load logic for special cases such as the Chessboard Puzzle or Montreaux's Office.
The general save/load logic in my last commit works in many cases, but there are certain special scenes where more is needed. For example, the laser heads in Montreaux's Office store some additional state data that needs to be saved and loaded. Or the Chessboard Room has a custom script that stores state that needs to be saved/loaded.