Updated Bridge Puzzle to better handle scene save states.
I'm trying to avoid saving/loading the animation system, since that would require save/loading callbacks, which just isn't a thing right now. To accomplish this with bridge tile glint/glow animations, I reworked the logic to be based on an update loop instead of using callbacks. This allows the states of the tiles to properly save/load.