Fix oversight where setting a GKActor active would always enable its shadow, even if it should be disabled.
The previous code would always enable the shadow when GKActor was set active, even if the shadow had previously been disabled. The oversight is that we need to track the shadow enabled state as a dedicated variable - we can't just assume that an active GKActor means an active shadow.

This fixes a bug where Gabe's shadow is errantly enabled on the altar during the Pendulum Puzzle.