Fixed several instances where characters face the wrong way before performing an animation.
This was a tricky one to fix, but it seems to work well. In short, the majority of 3D models in the game look down the negative Z axis by default. But the characters look down positive Z by default. Why the discrepancy? Who knows - I guess a miscommunication between programmers and art! The solution generally is to add PI to a model's rotation to get the actor's rotation.

However, there were a few instances where this isn't the case. For example, Mosely taking a mint in the lobby, or Emilio opening the hotel door during Day 3, 6PM. So it was inconsistent. Most models looked down negative Z, but some didn't.

To resolve that, this commit introduces some code that detects if a model is or isn't facing down negative Z and acts accordingly.