Added simple WIP "raycast diagnostic tool."
There are enough little raycast bugs that occur that I thought it'd be helpful to have a way to see what the raycast algorithms were detecting while the game was running. This commit adds a simple IMGUI tool that displays raycasts collisions that occurred on the current frame, along with their "t" values.

May or may not be useful long term, but I do intend to use it to fix a few outstanding raycast-related bugs.