Fixed dialogue camera not always being used when starting a conversation.
I noticed at least one instance (talking to Marcie in the train station) where it seemed odd that the camera didn't cut to the two characters when they were talking. On closer inspection, this is due to slightly incorrect logic when choosing a dialogue camera when entering a conversation.

When I observed the original game, it will use the "final" dialogue camera for the initial one IF no initial dialogue camera exists. I updated the logic slightly and it now works as it did in the original game.