Fixed a bug where actions executed by background timers would errantly cause the action bar to be canceled.
When an action is executed, we were forcing the action bar closed, since that usually means a cutscene is occurring. But an oversight was actions that occur in the background and aren't meant to be noticed by the player, such as ones that occur when a game timer expires. In this case, the action should execute, but the action bar should remain up.

This change provides a way to distinguish between user-initiated actions and background actions. The logic can then take different routes in the two cases.

Also, the TRUE condition for hiding the action bar is if the loading icon appears, which happens if the action doesn't resolve right away (there are waitable components in the action script). So, the logic for hiding the action bar is no longer simply "an action is starting" - it's "an action is currently playing, hide myself" instead.