Fixed a bug where skipping a video before a timeblock screen could errantly cause the timeblock screen to also be skipped.
Videos use ESC as a skip shortcut, and the timeblock screen also uses ESC to dismiss the screen and continue the game. A problem that could occur, when simply checking if ESC is pressed, is that the video update AND the timeblock screen update could run on the same frame, resulting in both screens "eating" the ESC input. The result for the player would be confusing: the video is skipped and the timeblock screen immediately dismisses itself.

This same sort of problem exists on the Finished Screen, so rather than duplicate the code, I tried to create a shared class that encompasses the logic to solve the problem. On the Finish Screen and Timeblock Screen, rather than just checking ESC, we also need to make sure that we detect at least one frame with no inputs. This allows us to be sure that the input we detected was a "fresh" or "new" input that should be acted upon.

There are other ways to solve this problem of course (an event-based input system, marking inputs as used, etc). But this was a decent approach that avoids adding a large new system right now!