Fixed logic for talking to Mosely about "whodunnit" when investigating the Devil's Armchair scene.
This scene uses a few fairly custom bits of logic. It's the only instance where a topic has type "recurring topic" instead of just "topic." It's the only place in the game where a custom CANCEL action is used. Besides the handshake puzzle, it's the only time a topic is meant to appear multiple times.