Implemented animation of mesh AABBs.
It was an oversight that this was never implemented. For the vast majority of the game, it also doesn't make much of a difference. But there were a few key points where I noticed that AABBs were not matching the original game. One particularly buggy one was the closet door in the attic at Chateau de Serres. Implementing this fixes the rare instances where an object's AABB changes as part of it's animation.