Changed action skip tick duration from 1 second to 0.1 seconds.
This is a tricky tweak with some tradeoffs. The problem ultimately is that some animation/walk code in the game doesn't work well with large delta times. If you action skip with a large delta time, some cutscenes resolve with characters in the wrong positions or facing the wrong direction. A stopgap solution is to reduce the delta time, though this means many more ticks will occur (potentially more computationally expensive). Action skips may not be as quick/responsive as a result.

Long term, the better way to resolve this is to improve the actor positioning and animation code. But for now, this helps resolve some bugs that can occur with action skip and actor animations/positioning.