Streamlined BSPActor creation a bit. Added visibility/interactivity tracking flags to BSPActor, which helps fix a save/load bug that can occur in some instances.
The impetus for this commit was a save/load bug: if you took the wire hanger from the closet in R25, created a save, and then loaded it, the wire hanger's hit test would be errantly visible in the closet. The cause is fairly complex - the hit test object actually changes from a hit test to a scene model after the wire hanger is taken, AND there's some complex interplay between when a hit-test vs. a scene-model uses SetActive vs SetVisible/SetInteractive.

ANYWAY - the crux of the issue is that BSPActor doesn't keep track of whether it was told to be visible or interactive. When you set it active, it always just assumes it should be both visible and interactive. This isn't a great assumption to make, since we sometimes have explicitly disabled visibility or interactivity. So, for example, if you said: SetVisible(false), SetActive(false), SetActive(true), the object would automatically be visible again.

So in addition to fixing the bug at hand, I think this commit also makes the BSPActor class a bit more logically sound. First, it now keeps track of visible and interactive state variables, which reflect the underlying BSP. When inactive, both these are overridden to false; but when set active, we still honor those flags. So setting active will not automatically set visible/interactive to true anymore.

Second, at key moments on creation, before becoming inactive, and on load, BSPActor refreshes these flags based on the actual state of the BSP itself. This just helps to avoid the BSPActor variables being out of date due to direct modifications to the BSP itself.

My hope is that this will result in overall better results in addition to fixing a bug. But more in depth testing is needed!