Removed "async" versions of AssetManager Load functions.
The goal of the async AssetManager functions was to speed up initial load times. A few managers that load a lot of assets on game launch (VerbManager, FootstepManager) could offload the time-consuming loading to a background thread. However, as implemented, it's a bit inelegant (requiring separate "Async" functions for each asset type). It also didn't work well on Linux.

Instead, what I want to try here is to instead manually wrap long asset loads in Loader::Load blocks (which is what the Async versions of these functions were doing anyways). This should allow achieving the same goal (fast launch times) while avoiding complicating the AssetManager API.

This is going on a feature branch for starters because it needs to be tested on all platforms before committing to it.