WIP converting AssetManager to not care about specific asset types.
A goal here is to make AssetManager game-agnostic - in which case it shouldn't be hardcoded to work with game-specific assets.
This commit gets closer to that goal by eschewing most public AssetManager functions used to load particular asset types.
Generic load commands are now used.

There's still work to be done here, but I think this is overall a good track.