Consolidated class hierarchy for input/output stream reading/writing in both binary/text.
This commit wrangles several existing and disparate classes (BinaryReader, BinaryWriter, TextWriter) into having shared base classes (StreamReader and StreamWriter). The idea here is to consolidate some shared functionality and ensure the public interfaces are consistent.

This also adds a previously missing "TextReader" class which complements TextWriter/BinaryReader.

Finally, all reader/writer classes now have the option of passing in a stream that is not owned by the object (rather than creating their own streams). Though it's often convenient for the reader/writer to own its stream, I foresee some upcoming work where it will be convenient to decouple the lifetime of a stream from the lifetime of a reader/writer. With this change, both approaches are now possible.