Moved Vector2/Vector3/Vector4 string parsing code directly into those classes. Fixed DrivingScreen needing to use an IniReader to read in path data, even though the path data wasn't in INI format.
The vast majority of GK3 data files are INI format. But at least one (PATHDATA.TXT for the driving screen) was not. For expediency, I previously "hacked it" such that this data was read in with the IniReader. So, this is just circling back and resolving that. Based on recent changes, a TextReader can now be used.

Also, I was creating temporary IniKeyValuePair instances because the only vector parsing code was baked into this class. That seemed wrong to me, and I instead moved that code directly to Vector2/3/4. This now allows parsing strings in a pretty robust/resilient way even if not using the IniReader class.